moving form to userDetails form and manage states from tab component
